Preuba de db-data 2

+++ controller/order.js
@@ -55,11 +55,11 @@ module.exports = {
 
         try {
             if (Object.keys(req.body).length === 0) {
-                return res.status(400).send('No hay productos en la orden');
+                return res.status(400).send();
 
             };
             if (!products || products.length === 0) {
-                return res.status(400).send();
+                return res.status(400).send('No hay productos en la orden');
             };
             const newOrder = new Order({
                 userId,
@@ -84,7 +84,7 @@ module.exports = {
     // //PUT "Modificar orden" -'/orders/:orderId'  
 
     updateOrder: async (req, res,next) => {
-        const {orderId, status}= req.body;
+        // const {orderId, status}= req.body;
         // if(req.userauth.roles.admin === false) {
         //     res.sendStatus(403)
         //  
@@ -93,12 +93,13 @@ module.exports = {
         //     return next(400);
         //   }
 
-        
+         const {orderId} = req.params.orderId;
+        const {status} = req.body;
         try{
             
-            const orderId = req.params.orderId;
+           
             if (Object.keys(req.body).length === 0) {
-                return res.status(404).send('No hay orden');
+                return res.status(400).send('');
                 
             };
             if (!isObjectId(orderId)) return next(400);
@@ -110,7 +111,6 @@ module.exports = {
               ];
             if (status && !statusOrder.includes(status)) return next(400);
             console.log(1, orderId);
-            
         const orderUpdate = await Order.findOneAndUpdate(
             {_id: orderId},
             {$set: req.body},
@@ -127,17 +127,21 @@ module.exports = {
 
     deleteOrder: async (req, res, next) => {
 
-
-
         try {
+        
             const orderId = req.params.orderId;
             console.log(orderId);
             if (!req.headers.validated) {
                 return res.status(401).send('No esta autenticado')
             };
+            
             if (!isObjectId(orderId)) {
                 return res.status(404).send('No se encuentra orden');
             };
+            if (!req.headers.validate.roles.admin){
+                return res.status(403).send();
+            }
+            
             const findOrder = await Order.findOne({ _id: orderId });
 
             await Order.findOneAndDelete({ _id: orderId });
